Move runtime parameters to ptroperty_tree
This is an API breaking commit!

Parameters for runtime classes were moved to general parameter structure
(strored in boost::property_tree). So the following code:

    boost::property_tree::ptree prm;
    typedef amgcl::runtime::make_solver< amgcl::backend::builtin<double> > Solver;
    Solver S(
        amgcl::runtime::coarsening::smoothed_aggregation,
        amgcl::runtime::relaxation::spai0,
        amgcl::runtime::solver::bicgstab,
        A,
        prm
        );

would have to be changed to

    boost::property_tree::ptree prm;
    typedef amgcl::runtime::make_solver< amgcl::backend::builtin<double> > Solver;
    prm.put("amg.coarsening.type", amgcl::runtime::coarsening::smoothed_aggregation);
    prm.put("amg.relaxation.type", amgcl::runtime::relaxation::spai0);
    prm.put("amg.coarsening.type", amgcl::runtime::solver::bicgstab);
    Solver S(A, prm);

This is done as part of the effort to generalize use of preconditioners
in the library, so that any preconditioner supported by AMGCL could be
constructed with matrix and parameters arguments.