Improve performance of relaxation::as_preconditioner
This gets rid of some unnecessary operations in as_preconditioner::apply().
This also reduces number of smoother applications to a neccessary
minumum. That is, single sweep is done in most cases, except for
gauss-seidel, where forward and backward sweeps are done to keep
preconditioner symmetry. This will increase the total number of
iterations, but will still be cheaper in terms of compute time.

Each relaxation method now provides apply_pre(), apply_post(), and
apply() methods. The former two are used inside V-cycle, as before, and
the latter is used as preconditioner.