Revert to packing iterative solver with mpi::subdomain_deflation
Subdomain deflation is too closely tied with the iterative solver
to be used as a normal preconditioner. For example, it requires
postprocessing step after the solution is finished, and its not
possible to apply subdomain deflation to a matrix different from
the one it has been setup for.

With this change, the subdomain deflation may be used as a
normal preconditioner again.