(Optional) asynchronous amg setup
amgcl::amg constructor starts the setup process in a new thread. As soon
as constructor returns, the instance is ready to be used as a
preconditioner. Initially its just a sigle-level smoother, but when as
the new (coarser) levels are constructed, they are put to use.

In case of GPGPU backends, this should allow to overlap work between
host CPU doing setup and the compute device doing the solution. In some
cases a 2x speedup of the overall solution has been achieved.