Fully resolved solver class in template parameter of mpi::make_solver
That is, explicitly specify iterative solver backend instead of
inheriting the backend from the preconditioner. This is a first step
towards using mixed precision in mpi context.

Also provide convenience wrappers for iterative solvers in mpi
namespace, so that users do not have to explicitly specify
mpi::inner_product in template parameters of the iterative solver.

Refs #137